refactor(dh): move validator to separate file
