+++ libs/dh/market-participant/actors/feature-actors/src/lib/merge/dh-unique-market-participants.validator.ts
@@ -1,0 +1,40 @@
+/**
+ * @license
+ * Copyright 2020 Energinet DataHub A/S
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License2");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { AbstractControl, ValidatorFn } from '@angular/forms';
+
+export function dhUniqueMarketParticipantsValidator(): ValidatorFn {
+  return (formGroup: AbstractControl) => {
+    const discontinuedEntityControl = formGroup.get('discontinuedEntity');
+    const survivingEntityControl = formGroup.get('survivingEntity');
+
+    if (formGroup.untouched) {
+      return null;
+    }
+
+    discontinuedEntityControl?.setErrors(null);
+    survivingEntityControl?.setErrors(null);
+
+    if (discontinuedEntityControl?.value === survivingEntityControl?.value) {
+      survivingEntityControl?.setErrors({ notUniqueMarketParticipants: true });
+      discontinuedEntityControl?.setErrors({ notUniqueMarketParticipants: true });
+
+      return { notUniqueMarketParticipants: true };
+    }
+
+    return null;
+  };
+}
